refactor(providers): merge `JsonPubSubClient` into `JsonRpcClient`
